+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 - Do not crash if permission not granted for in phone call constraint and trying to perform key map.
 - #2258 remember saved position in key map list when changing them. Do not flash with loading indicator briefly as well.
+- #2167 do not show errors for app not in foreground constraint when the app is uninstalled/disabled.
 
 ## [4.4.1](https://github.com/sds100/KeyMapper/releases/tag/v4.4.1)